+++ libCommon/test/event/HandlerEventTest.cpp
@@ -37,9 +37,9 @@ int test::eventTest::HandlerEvents()
     EventHandler handler = {1};
     EventArgs args = {0, 0, 0};
 
-    assert::AreEqual(0u, event.Size());
+    assert::AreEqual(size_t(0), event.Size());
     event += handler;
-    assert::AreEqual(1u, event.Size());
+    assert::AreEqual(size_t(1), event.Size());
 
     event(args);
 
@@ -47,9 +47,9 @@ int test::eventTest::HandlerEvents()
     assert::AreEqual(RES_Y + handler.id, args.y);
     assert::AreEqual(RES_Z + handler.id, args.z);
 
-    assert::AreEqual(1u, event.Size());
+    assert::AreEqual(size_t(1), event.Size());
     event -= handler;
-    assert::AreEqual(0u, event.Size());
+    assert::AreEqual(size_t(0), event.Size());
 
     args = {0, 0, 0};
     event(args);
@@ -57,7 +57,7 @@ int test::eventTest::HandlerEvents()
     assert::AreEqual(0, args.x);
     assert::AreEqual(0, args.y);
     assert::AreEqual(0, args.z);
-    assert::AreEqual(0u, event.Size());
+    assert::AreEqual(size_t(0), event.Size());
 
     return 0;
 }
