+++ libCommon/test/event/MemberHandlerEventTest.cpp
@@ -33,9 +33,9 @@ int test::eventTest::MemberHandlerEvents()
     auto member_handler = util::MakeMemberHandler(&handler, &EventHandler::HandleSomeEvent);
     EventArgs args = {0, 0, 0};
 
-    assert::AreEqual(0u, event.Size());
+    assert::AreEqual(size_t(0), event.Size());
     event += member_handler;
-    assert::AreEqual(1u, event.Size());
+    assert::AreEqual(size_t(1), event.Size());
 
     event(args);
 
@@ -43,9 +43,9 @@ int test::eventTest::MemberHandlerEvents()
     assert::AreEqual(RES_Y + handler.id, args.y);
     assert::AreEqual(RES_Z + handler.id, args.z);
 
-    assert::AreEqual(1u, event.Size());
+    assert::AreEqual(size_t(1), event.Size());
     event -= member_handler;
-    assert::AreEqual(0u, event.Size());
+    assert::AreEqual(size_t(0), event.Size());
 
     args = {0, 0, 0};
     event(args);
@@ -53,7 +53,7 @@ int test::eventTest::MemberHandlerEvents()
     assert::AreEqual(0, args.x);
     assert::AreEqual(0, args.y);
     assert::AreEqual(0, args.z);
-    assert::AreEqual(0u, event.Size());
+    assert::AreEqual(size_t(0), event.Size());
 
     return 0;
 }
